fix(jwt): keep key entry assertion internal
