+++ packages/jwt/src/signing/signer.ts
@@ -3,7 +3,8 @@ import { Inject } from '@fluojs/core';
 import { JwtConfigurationError } from '../errors.js';
 import { normalizeRefreshTokenOptions } from '../refresh/refresh-token.js';
 import type { JwtAlgorithm, JwtClaims, JwtKeyEntry, JwtVerifierOptions } from '../types.js';
-import { ASYMMETRIC_HASH, assertJwtKeyEntries, HMAC_HASH, JWT_OPTIONS } from './verifier.js';
+import { assertJwtKeyEntries } from './key-entries.js';
+import { ASYMMETRIC_HASH, HMAC_HASH, JWT_OPTIONS } from './verifier.js';
 
 function encodeBase64Url(value: Buffer | string): string {
   return Buffer.from(value)
